+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fix issue with missing ignores in prettier template conf
+- Run ruff linting in project CI
 - Upgrade next-i18next to 13.0.1
 - Upgrade prettier to 2.8.2
 - Upgrade django-stubs to 1.14.0